add blogger endpoint (remove blogs)

+++ src/blogger/blogs/use-cases/remove.blog.use.case.ts
@@ -1,10 +1,11 @@
 import { CommandHandler } from '@nestjs/cqrs';
 import { BlogsRepository } from '../../../blogs/blogs.repository';
 import { DbId } from '../../../global-types/global.types';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import { CurrentUserType } from '../../../auth/types/current.user.type';
 
 export class RemoveBlogCommand {
-  constructor(public id: DbId) {}
+  constructor(public id: DbId, public user: CurrentUserType) {}
 }
 
 @CommandHandler(RemoveBlogCommand)
@@ -14,6 +15,8 @@ export class RemoveBlogUseCase {
   async execute(command: RemoveBlogCommand): Promise<boolean> {
     const blog = await this.blogsRepository.getById(command.id);
     if (!blog) throw new NotFoundException();
+    if (blog.blogOwnerInfo.userId !== command.user.userId)
+      throw new ForbiddenException();
     await this.blogsRepository.delete(command.id);
     return true;
   }
